% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellTagForCollapsing.R
\name{CellTagDataForCollapsing}
\alias{CellTagDataForCollapsing}
\title{CellTag Starcode Prior Collapsing}
\usage{
CellTagDataForCollapsing(ctm.after.whitelist, umi.matrix, output.file)
}
\arguments{
\item{ctm.after.whitelist}{CellTag single-cell matrix after whitelist filtering}

\item{umi.matrix}{CellTag raw UMI count matrix}

\item{output.file}{The filepath and name to save the table for collapsing}
}
\value{
The data frame for collapsing
}
\description{
This function generate the .txt file that will be fed into starcode - https://github.com/gui11aume/starcode - to collapse similar CellTags.
}
\examples{
CellTagDataForCollapsing(celltags.whitelisted.3, combo.umi.mat.10, "my_favoriate_collapse.csv")

}
\keyword{CellTagging}
\keyword{RNA-seq}
\keyword{data,}
\keyword{single-cell}
