% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellTagForCollapsing.R
\name{CellTagDataPostCollapsing}
\alias{CellTagDataPostCollapsing}
\title{CellTag Starcode Post Collapsing}
\usage{
CellTagDataPostCollapsing(celltag.obj, collapsed.rslt.file)
}
\arguments{
\item{collapsed.rslt.file}{File path to the collapsed result file}

\item{ctm.after.whitelist}{CellTag single-cell matrix after whitelist filtering}

\item{collapsed.csv.file}{File path to the data frame file generated for collapsing}

\item{output.file}{The RDS file path and name to save the resulting UMI matrix}
}
\value{
The collapsed and processed UMI matrices
}
\description{
This function processes the result generated from starcode - https://github.com/gui11aume/starcode.
}
\examples{
CellTagDataPostCollapsing(celltags.whitelisted.3, "collapsed_test.txt", "collapsed.csv", "collapsed_data_matrix.Rds")

}
\keyword{CellTagging}
\keyword{RNA-seq}
\keyword{data,}
\keyword{single-cell}
