% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScCellTagMatrixProcess.R
\name{SingleCellDataBinatization}
\alias{SingleCellDataBinatization}
\title{Single-cell RNA-seq Binarization Function}
\usage{
SingleCellDataBinatization(celltag.obj, tag.cutoff)
}
\arguments{
\item{celltag.obj}{A CellTag object with the raw count matrix generated}

\item{tag.cutoff}{How many tags would you like to be used as a cutoff to say that the cells are tagged?}
}
\value{
A CellTag object with the attribute (binary.mtx) filled.
}
\description{
This function binarize the single-cell celltag data
}
\examples{
SingleCellDataBinatization(bam.test.obj, 2)

}
\keyword{CellTagging}
\keyword{RNA-seq}
\keyword{data,}
\keyword{single-cell}
