% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScCellTagMatrixProcess.R
\name{SingleCellDataWhitelist}
\alias{SingleCellDataWhitelist}
\title{Single-cell RNA-seq Whitelisting Function}
\usage{
SingleCellDataWhitelist(celltag.obj, whitels.cell.tag.file)
}
\arguments{
\item{celltag.obj}{A CellTag object with the binary matrix generated}

\item{whitels.cell.tag.file}{file director to the whitelisted cell tags}
}
\value{
A CellTag object with the attribute (whitelisted.count) filled
}
\description{
This function conducts whitelist filtering through the single-cell dataset
}
\examples{
SingleCellDataWhitelist(bam.test.obj, "~/Desktop/My_Favourite_Whitelist.csv")

}
\keyword{CellTagging}
\keyword{RNA-seq}
\keyword{data,}
\keyword{single-cell}
